Se ha avanzado en la practica, tiens la funcion de agregar eliminar actualizar y ver cursos, tiene la funcion de agregar actualiza y eliminar la cuenta (user) y tiene las validaciones de no poder actualizar o eliminar cuentas externas, no puede eliminar agregar o actualizar, si posee el rol de STUDENT_ROL

+++ Practicas Supervisadas/Practica Supervisada/src/Curso/Curso.controller.js
@@ -9,8 +9,8 @@ export const testCuso = (req,res)=>{
 export const addCurso = async(req, res)=>{
     try{
         let data = req.body
-        let user = await User.findOne({_id:data.user})
-        if(!user) return res.status(404).send({message: 'User invalid | id user'})
+        let {  role } = req.user
+        if(role == 'STUDENT_ROLE') return res.status(403).send({message: 'You do not have authorization to add courses'})
         let curso = new Curso(data)
         await curso.save()
         return res.send({message: 'Course successfully added'})
@@ -20,28 +20,10 @@ export const addCurso = async(req, res)=>{
     }
 }
 
-export const viewCurso = async(req, res)=>{
-    try{
-        let { name } = req.body
-        let curso = await Curso.findOne({name})
-        if (curso){
-            let loggedCurso = {
-                name: curso.name,
-                grade: curso.grade,
-                description: curso.description
-            }
-            return res.send({message: `The ${product.name} found`, loggedProduct})
-        }
-        return res.status(404).send({message: 'Invalid Name'})
-    }catch(error){
-        console.error(err)
-        return res.status(404).send({message: 'Error when searching'})
-    }
-}
-
 export const lookForAllCursos = async(req, res)=>{
     try{
-        let all = await Curso.find({})
+        let { uid } = req.user
+        let all = await Curso.find({teacher:uid})
         return res.send({message: all})
     }catch(err){
         console.error(err)
@@ -52,6 +34,9 @@ export const lookForAllCursos = async(req, res)=>{
 export const updateCurso = async(req, res)=>{
     try{
        let { id } = req.params
+       let cursoData = await Curso.findOne({_id:id})
+       let { uid } = req.user
+       if(cursoData.teacher != uid) return res.status(403).send({message: 'You do not have authorization to modify the course'})
        let data = req.body
        let updatedCurso = await Curso.findOneAndUpdate(
         {_id: id},
@@ -68,11 +53,51 @@ export const updateCurso = async(req, res)=>{
 export const deleteCurso = async(req,res)=>{
     try{
         let { id } = req.params
+        let cursoData = await Curso.findOne({_id:id})
+        if(!cursoData) return res.status(404).send({message: 'Course not found'})
+        if(cursoData.teacher != uid) return res.status(403).send({message: 'You do not have authorization to modify the course'})
         let deleteCur = await Curso.findOneAndDelete({_id: id})
         if(!deleteCur) return res.status(404).send({message: 'The course could not be deleted'})
         return res.send({message: `The course: ${deletedProduct.name} has been successfully removed`})
-    }catch(error){
+    }catch(err){
         console.error(err)
         return res.status(404).send({message: 'Unexpected error while deleting'})
+    }
+}
+/*
+export const unirseCurso = async(req,res)=>{
+   
+        if(cursos.length < 3){
+            let cursoUnir = await Curso.find({_id:id})
+            if (!cursoUnir) return res.status(404).send({ message: 'Curso not found' });
+            if (cursoUnir.students.includes(uid)) return res.status(400).send({ message: 'Student is already enrolled in this course' });
+            cursoUnir.students.push(uid)
+            return res.send({message: 'Unido'})
+        }
+}*/
+
+export const unirseCurso = async (req, res) => {
+    try {
+        let { id } = req.params;
+        let { uid } = req.user;
+        let cursos = await Curso.find({ students: uid });
+        if (!cursos) {
+            return res.status(404).send({ message: 'Course not found' });
+        }
+        if (cursos.students && cursos.students.includes(uid)) {
+            return res.status(400).send({ message: 'You are already enrolled in this course' });
+        }
+        if (cursos.length < 3) {
+            let updateCurs = await Curso.findOneAndUpdate(
+                { _id: id },
+                { students: uid }
+            );
+            if(!updateCurs) return res.status(400).send({message: 'Login error'})
+            return res.send({ message: 'Unido' });
+        }
+        return res.send({ message: 'It is not possible to join more courses' });
+    } catch (err) {
+        console.error(err);
+        return res.status(404).send({ message: 'Unexpected error when adding' });
     }
 }
